+++ go_back/cmd/server/main.go
@@ -15,8 +15,8 @@ import (
 
 	"gng/internal/database/postgres"
 	"gng/internal/handlers"
-	"gng/internal/helper"
 	"gng/internal/services"
+	"gng/internal/utils/helpers"
 	"gng/internal/utils/logger"
 )
 
@@ -24,7 +24,7 @@ func main() {
 	// Инициализация логгера
 	log := logger.NewLogger()
 
-	if err := godotenv.Load(helper.GetEnv("ENV_PATH", ".env")); err != nil {
+	if err := godotenv.Load(helpers.GetEnv("ENV_PATH", ".env")); err != nil {
 		log.Info("No .env file found, using system environment variables")
 	}
 
@@ -73,10 +73,7 @@ func main() {
 		w.Write([]byte("OK"))
 	})
 
-	port := os.Getenv("PORT")
-	if port == "" {
-		port = "8080"
-	}
+	port := helpers.GetEnv("PORT", "8080")
 
 	srv := &http.Server{
 		Addr:         ":" + port,
