+++ go_back/internal/utils/helpers/env.go
@@ -1,4 +1,4 @@
-package helper
+package helpers
 
 import (
 	"os"
